splitting up file system sync service
