+++ src/Service/Sync/Interfaces/FileServiceInterface.php
@@ -1,0 +1,27 @@
+<?php
+
+/*
+ * This file is part of the mangel.io project.
+ *
+ * (c) Florian Moser <git@famoser.ch>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Service\Sync\Interfaces;
+
+use App\Entity\Traits\FileTrait;
+
+interface FileServiceInterface
+{
+    /**
+     * @param string $folder
+     * @param string $ending
+     * @param FileTrait[] $knownFiles
+     * @param callable $createNewFile
+     *
+     * @return FileTrait[]
+     */
+    public function getFiles(string $folder, string $ending, array $knownFiles, callable $createNewFile);
+}
